Revert "Refactor country information:"
This reverts commit 0b5f753fc33f9ca1dcd6660f92ee3eea17a9fba0.
