+++ src/domain/entities/GlassModule.ts
@@ -11,7 +11,6 @@ export interface GlassModule {
     id: string;
     userGroups: ModuleUserGroups;
     questionnaires: QuestionnaireConfig[];
-    countryInformationId?: Id;
 }
 
 interface QuestionnaireConfig {
